feat: usersテーブル追加、各ページにユーザーと不正入力のチェックを追加

+++ WEB-INF/source/jp/co/wordbook/controllers/QuizDeleteServlet.java
@@ -13,11 +13,33 @@ public class QuizDeleteServlet extends HttpServlet {
     public void doGet(HttpServletRequest request, HttpServletResponse response)
         throws ServletException, IOException
     {
-        int quiz_id = Integer.parseInt(request.getParameter("id"));
+        final int IRREG = -1;
+        
+        // セッション、パラメータから取得
+        String userId = Session.getUserId(request);
+        int quiz_id = NoNull.parseInt(request.getParameter("id"), IRREG);
 
-        // データベースから取得、削除
+        // 不正な入力 → インフォメーションページへ
+        if (quiz_id == IRREG) {
+            Information.forwardDataWasIncorrect(request, response);
+            return;
+        }
+
+        // データベースから取得
         QuizDAO quizDAO = new QuizDAO();
         QuizBean quiz = quizDAO.getRecord(quiz_id);
+
+
+        // 問題が取得できない、ユーザーと紐づいていない 
+        //  → 不正入力のインフォメーションページへ
+        if (quiz == null ||
+            !new SubjectDAO().userHasSubject(quiz.getSubject_id(), userId)
+        ) {
+            Information.forwardDataWasIncorrect(request, response);
+            return;
+        }
+
+        // データベースから削除
         boolean success = quizDAO.deleteRecord(quiz_id);
         
         // 見出し
@@ -25,15 +47,14 @@ public class QuizDeleteServlet extends HttpServlet {
             ? "問題を削除しました"
             : "問題を削除できませんでした";
 
-        // リクエストへ設定
-        request.setAttribute("heading", heading);
-        request.setAttribute("paragraph", "問題文 : " + quiz.getQuestion());
-        request.setAttribute("buttonname", "問題一覧へ");
-        request.setAttribute("url", "quizlist?subjectid=" + quiz.getSubject_id());
 
-        // JSPへ送信
-        String view = "/WEB-INF/views/information.jsp";
-        RequestDispatcher dispatcher = request.getRequestDispatcher(view);
-        dispatcher.forward(request, response);
+        // infomation.jsp へ送信
+        Information.forward(
+            request, response, 
+            heading,
+            "問題文 : " + quiz.getQuestion(),
+            "問題一覧へ",
+            "quizlist?subjectid=" + quiz.getSubject_id()
+        );
     }
 }
